Establish consistent LearningTargetsRepo name
1. Java convention discourages using multiple words for package names.
If absolutely necessary, and underscore can be used to separate words.
Camel case is not allowed by standard convention. Thus, this patch renames
the 'learningTargets' package to 'targets'.

2. We have been using the plural for our repositories (GoalsRepo and
TasksRepo not GoalRepo and TaskRepo). To follow this convention, this patch
renames the LearningTargetRepo to LearningTargetsRepo. All implementing
classes also have their names adjusted accordingly.